Changes needed to make it build after increasing compiler warnings.

+++ src/libs/test-libs/utillib/cmdLineTest/cmdLineTest.cpp
@@ -39,7 +39,7 @@ TEST_F(ThisTest, Test1) {
     if (opts.testNum ==  5) ASSERT_TRUE( "5. int  == 12",           (opts.numOption ==   12) );
     if (opts.testNum ==  6) ASSERT_TRUE( "6. int  == -100",         (opts.numOption == -100) );
     if (opts.testNum ==  7) ASSERT_TRUE( "7. uint == 12",           (opts.numOption ==   12) );
-    if (opts.testNum ==  8) ASSERT_TRUE( "8. uint == -100 (false)", (opts.numOption == NOPOS) );
+    if (opts.testNum ==  8) ASSERT_TRUE( "8. uint == -100 (false)", (opts.numOption == (int64_t)NOPOS) );
     if (opts.testNum ==  9) ASSERT_TRUE( "9. string with string",   (opts.stringOption == "This is a string") );
     if (opts.testNum == 10) ASSERT_TRUE("10. string with 1212",     (opts.stringOption == "1212") );
